requirements.txt
pyfrc
wpilib
marsutils
pynetworktables
robotpy-ctre
robotpy-navx
robotpy-rev
pygame